Upgraded projects to vs2017 csproj format and netstandard/netcore, fixed unit tests, use sbyte rather than byte since usage expects a signed 8bit integer, fixed generic interger conversion always decoding to int64

+++ rencode-sharp/BStruct.cs
@@ -41,11 +41,11 @@ namespace rencodesharp
 		/// <summary>
 		/// Converts byte array to INT1 (8 bit integer)
 		/// </summary>
-		public static byte ToByte(byte[] value, int startIndex)
+		public static sbyte ToByte(byte[] value, int startIndex)
         {
             if (value.Length != 1)
                 throw new ArgumentException("\"value\" doesn't have 1 byte.");
-		    return value[0];
+		    return (sbyte)value[0];
         }
 
 		/// <summary>
